requirements.txt
httpx==0.24.1
openai==0.27.8
Requests==2.31.0
websocket_client==1.6.1
